Fix T74256, crash in bmesh.ops.bevel if segments not set.
Also changed signature of bevel function to take integer
for segments instead of float, which is just wrong.

+++ source/blender/bmesh/tools/bmesh_bevel.h
@@ -27,7 +27,7 @@ struct MDeformVert;
 void BM_mesh_bevel(BMesh *bm,
                    const float offset,
                    const int offset_type,
-                   const float segments,
+                   const int segments,
                    const float profile,
                    const bool vertex_only,
                    const bool use_weights,
